refactor: Lotto Bounded Context 변경
- 구매 BC에서 shared kernel로 이동
- 패키지 이동
